Add subquery support and join builder enhancements
- Implemented SubQueryTest for testing lambda-based subqueries in HQL.
- Added JoinBuilder class for configuring joins with inline conditions and nested joins.
- Created SubQueryContext for building subqueries with a fluent API.
- Developed WhereGroupContext for building grouped WHERE conditions with AND/OR logic.
- Enhanced existing builder methods to support nested queries and conditions.

+++ hibernate-criteria-builder/src/test/java/io/github/robertomike/hefesto/hefesto/constructors/SelectBuilderTest.java
@@ -66,7 +66,7 @@ public class SelectBuilderTest {
     void selectAvgIdOfUsers() {
         var result = new Hefesto<>(User.class)
                 .addSelect("id", SelectOperator.AVG)
-                .findFirstFor(Float.class);
+                .findFirstFor(Double.class);
 
         assertNotNull(result);
         assertTrue(result > 1);
